+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/video/VideoRenderView.swift
@@ -7,10 +7,11 @@
 //
 
 import Foundation
+import VideoToolbox
 
 /// `VideoRenderView` renders frame that comes from `VideoTile`.
 @objc public protocol VideoRenderView {
     /// Render given frame to UI
     /// - Parameter frame: a frame of video
-    func renderFrame(frame: Any?)
+    func renderFrame(frame: CVPixelBuffer?)
 }